Move lib/client/BaseSubscriptionGraphqlCapsule.js below lib/client/graphql/
