+++ lib/client/graphql/BaseSubscriptionGraphqlCapsule.js
